後端關於score部分由double改為long

See merge request tio/3n1-android!136

+++ lib_database/src/main/java/arch/cayenne/lib/database/entity/UserDataBean.kt
@@ -33,6 +33,6 @@ data class AvatarEmbedded(
 
 data class WalletBean(
     val currency: String,
-    val balance: Double,
+    val balance: Long,
     val convertedAmount: Long
 )
